+++ src/abacus/detail/raw.hpp
@@ -11,7 +11,6 @@
 #include <cstring>
 
 #include "../version.hpp"
-#include <bourne/json.hpp>
 
 namespace abacus
 {
@@ -23,7 +22,7 @@ namespace detail
 /// @return The size of the header in bytes
 inline auto header_bytes() -> std::size_t
 {
-    return 7;
+    return 5;
 }
 
 /// @param data The raw memory for the counters
@@ -40,31 +39,10 @@ inline auto max_name_bytes(const uint8_t* data) -> uint16_t
     return max_name_bytes;
 }
 
-/// @return The offset of the max_scope_bytes data in the header
-inline auto max_scope_bytes_offset() -> std::size_t
-{
-    return 2;
-}
-
-/// @param data The raw memory for the counters
-/// @return The maximum bytes a scope can contain
-inline auto max_scope_bytes(const uint8_t* data) -> uint16_t
-{
-    assert(data != nullptr);
-
-    const uint8_t* max_scope_bytes_data = data + max_scope_bytes_offset();
-    uint16_t max_scope_bytes;
-    std::memcpy(&max_scope_bytes, max_scope_bytes_data, sizeof(uint16_t));
-
-    assert(max_scope_bytes > 0);
-
-    return max_scope_bytes;
-}
-
 /// @return The max_metrics offset in the raw memory
 inline auto max_metrics_offset() -> std::size_t
 {
-    return 4;
+    return 2;
 }
 
 /// @param data The raw memory for the counters
@@ -82,18 +60,12 @@ inline auto max_metrics(const uint8_t* data) -> uint16_t
     return max_metrics;
 }
 
-/// @return The scope offset in the raw memory
-inline auto scope_offset() -> std::size_t
-{
-    return header_bytes();
-}
-
 /// @param data The raw memory for the counters
 /// @return The maximum metrics the raw memory can contain
-inline auto names_offset(const uint8_t* data) -> std::size_t
+inline auto names_offset() -> std::size_t
 {
-    // Skip header + scope
-    return header_bytes() + max_scope_bytes(data);
+    // Skip header
+    return header_bytes();
 }
 
 /// @param offset The offset in the raw memory
@@ -108,35 +80,13 @@ inline auto values_alignment_padding(std::size_t offset) -> std::size_t
 inline auto values_offset(const uint8_t* data) -> std::size_t
 {
     // Skip header + scope + names
-    std::size_t offset = header_bytes() + max_scope_bytes(data) +
-                         (max_metrics(data) * max_name_bytes(data));
+    std::size_t offset =
+        header_bytes() + (max_metrics(data) * max_name_bytes(data));
 
     // align to 8 bytes
     offset += values_alignment_padding(offset);
 
     return offset;
-}
-
-/// @param data The raw memory for the counters
-/// @return The raw scope in memory
-inline auto raw_scope(uint8_t* data) -> char*
-{
-    assert(data != nullptr);
-
-    uint8_t* scope_data = data + scope_offset();
-
-    return (char*)scope_data;
-}
-
-/// @param data The raw memory for the counters
-/// @return The raw scope in memory
-inline auto raw_scope(const uint8_t* data) -> const char*
-{
-    assert(data != nullptr);
-
-    const uint8_t* scope_data = data + scope_offset();
-
-    return (const char*)scope_data;
 }
 
 /// @param data The raw memory for the counters
@@ -147,8 +97,7 @@ inline auto raw_name(uint8_t* data, std::size_t index) -> char*
     assert(data != nullptr);
     assert(index < max_metrics(data));
 
-    uint8_t* name_data =
-        data + names_offset(data) + (index * max_name_bytes(data));
+    uint8_t* name_data = data + names_offset() + (index * max_name_bytes(data));
 
     return (char*)name_data;
 }
@@ -162,23 +111,9 @@ inline auto raw_name(const uint8_t* data, std::size_t index) -> const char*
     assert(index < max_metrics(data));
 
     const uint8_t* name_data =
-        data + names_offset(data) + (index * max_name_bytes(data));
+        data + names_offset() + (index * max_name_bytes(data));
 
     return (const char*)name_data;
-}
-
-inline void prepend_name(const std::string& text, std::size_t index,
-                         uint8_t* data)
-{
-    char* name_data = raw_name(data, index);
-    std::size_t text_size = text.size();
-    std::size_t name_size = std::strlen(name_data);
-    assert((text_size + 1 + name_size <= max_name_bytes(data)));
-
-    std::memcpy(name_data + text_size + 1, name_data, name_size);
-    std::memcpy(name_data, text.data(), text_size);
-    name_data[text_size] = '_';
-    // name_data[text_size + name_size + 1] = '\0';
 }
 
 /// @param data The raw memory for the counters
@@ -224,13 +159,37 @@ inline auto is_metric_initialized(const uint8_t* data, std::size_t index)
 }
 
 /// @param data The raw memory for the counters
-/// @return The counters in json-format
-inline auto to_json(const uint8_t* data) -> bourne::json
+/// @return The number of initialized metrics in memory
+inline auto metrics_count(const uint8_t* data) -> std::size_t
 {
     assert(data != nullptr);
-    bourne::json counters = bourne::json::object();
+
+    std::size_t count = 0U;
 
     for (std::size_t i = 0; i < max_metrics(data); ++i)
+    {
+        if (is_metric_initialized(data, i))
+        {
+            ++count;
+        }
+    }
+
+    return count;
+}
+
+/// @param data The raw memory for the counters
+/// @param prettier If true, the output will be more human-readable format.
+/// Otherwise, it will be compact JSON.
+/// @return The counters in json-format
+inline auto to_json(const uint8_t* data, bool prettier = false) -> std::string
+{
+    std::string space = prettier ? " " : "";
+    std::string newline = prettier ? "\n" : "";
+    std::string tab = prettier ? "\t" : "";
+    assert(data != nullptr);
+    std::string counters_json = "{" + newline;
+
+    for (std::size_t i = 0; i < metrics_count(data); ++i)
     {
         if ((!is_metric_initialized(data, i)))
         {
@@ -240,10 +199,16 @@ inline auto to_json(const uint8_t* data) -> bourne::json
         auto n = raw_name(data, i);
         auto v = *raw_value(data, i);
 
-        counters[n] = v;
+        counters_json +=
+            tab + "\"" + std::string(n) + "\":" + space + std::to_string(v);
+        if (i != (metrics_count(data) - 1U))
+        {
+            counters_json += "," + newline;
+        }
     }
+    counters_json += newline + "}";
 
-    return counters;
+    return counters_json;
 }
 
 }