+++ NEWS.rst
@@ -6,7 +6,16 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Major: metrics::initialize_metric() only take 1 argument "name".
+         Metrics are initialized with incrementally increasing index.
+* Major: Changed "title" argument to "scope".
+* Major: Added a scope member to metrics object which prefixes the names of
+         the metrics.
+* Minor: Added add_scope() member function to class metrics.
+* Minor: Added default argument for pretty-print for the to_json() member
+         functions. Default is false.
+* Minor: Removed title from the class metrics data-header.
+* Major: Removed the metrics::set_metrics_title() member funciton
 
 1.1.2
 -----